// Aplicacion/AppAdopciones/src/app/models/enums.ts
export enum TipoUsuario {
  ADOPTANTE = 'adoptante',
  ADMIN = 'admin'
}

export enum EstadoUsuario {
  ACTIVO = 'activo',
  INACTIVO = 'inactivo',
  BLOQUEADO = 'bloqueado'
}

export enum EstadoSolicitud {
  PENDIENTE = 'Pendiente',
  APROBADO = 'Aprobado',
  EN_REVISION = 'En revision',
  RECHAZADO = 'Rechazado',
  CANCELADO = 'Cancelado'
}

export enum EstadoMascota {
  ENFERMO = 'Enfermo',
  EUTANASIA = 'Eutanasia',
  ADOPTADO = 'Adoptado',
  EN_VETERINARIA = 'En veterinaria',
  FALLECIDO = 'Fallecido',
  DISPONIBLE = 'Disponible'
}

export enum EstadoAdopcion {
  PENDIENTE = 'Pendiente',
  APROBADO = 'Aprobado',
  EN_REVISION = 'En revision',
  RECHAZADO = 'Rechazado',
  CANCELADO = 'Cancelado'
}

export enum TipoMascotaEnum {
  PERRO = 'Perro',
  GATO = 'Gato',
  CONEJO = 'Conejo',
  AVE = 'Ave'
}

export enum EstadoSalud {
  ESTERILIZADO = 'Esterilizado',
  SIN_ESTERILIZAR = 'Sin Esterilizar'
}

export enum ColorChip {
  PRIMARY = 'primary',
  SECONDARY = 'secondary',
  TERTIARY = 'tertiary',
  SUCCESS = 'success',
  WARNING = 'warning',
  DANGER = 'danger',
  MEDIUM = 'medium',
  LIGHT = 'light',
  DARK = 'dark'
}

export enum IconName {
  CHECKMARK_CIRCLE = 'checkmark-circle',
  TIME = 'time',
  EYE = 'eye',
  CLOSE_CIRCLE = 'close-circle',
  BAN = 'ban',
  HELP_CIRCLE = 'help-circle',
  PERSON_CIRCLE = 'person-circle',
  PAW = 'paw',
  HEART = 'heart',
  HOME = 'home',
  MAIL = 'mail',
  CALL = 'call',
  LOCATION = 'location',
  BUSINESS = 'business',
  CALENDAR = 'calendar',
  SHIELD_CHECKMARK = 'shield-checkmark'
}